+++ tests/unit/aws_interface/test_models.py
@@ -1,4 +1,5 @@
-from cloudwanderer.aws_interface.models import IdPartSpecification
+from cloudwanderer.aws_interface.models import IdPartSpecification, RelationshipSpecification
+from cloudwanderer.models import RelationshipAccountIdSource, RelationshipDirection, RelationshipRegionSource
 
 
 def test_get_specified_urn_parts():
@@ -33,3 +34,28 @@ def test_get_urn_parts_no_regex():
     )
 
     assert subject.get_urn_parts(data) == {"resource_id_parts": ["subnet-111111111"]}
+
+
+def test_relationship_specification_factory():
+    subject = RelationshipSpecification.factory(
+        definition={
+            "basePath": "Id",
+            "idParts": "",
+            "direction": "outbound",
+            "service": "test_service",
+            "resourceType": "test_resource_type",
+            "regionSource": "sameAsResource",
+            "accountIdSource": "sameAsResource",
+        }
+    )
+
+    assert subject == RelationshipSpecification(
+        base_path="Id",
+        id_parts=[],
+        cloud_name="aws",
+        service="test_service",
+        resource_type="test_resource_type",
+        region_source=RelationshipRegionSource.SAME_AS_RESOURCE,
+        account_id_source=RelationshipAccountIdSource.SAME_AS_RESOURCE,
+        direction=RelationshipDirection.OUTBOUND,
+    )